img_seg/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.63.0
idna==3.7
##intel-openmp
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
##mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.1
pretty-errors==1.2.25
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
##pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
##tbb==2021.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
